+++ src/adapters/MeetingsJSONAdapter.js
@@ -14,6 +14,7 @@ export const JOIN_CONTROL = 'join-meeting';
 export const LEAVE_CONTROL = 'leave-meeting';
 export const DISABLED_MUTE_AUDIO_CONTROL = 'disabled-mute-audio';
 export const DISABLED_JOIN_CONTROL = 'disabled-join-meeting';
+export const ROSTER_CONTROL = 'member-roster';
 
 // TODO: Figure out how to import JS Doc definitions and remove duplication.
 /**
@@ -122,6 +123,12 @@ export default class MeetingsJSONAdapter extends MeetingsAdapter {
       action: () => {},
       display: this.disabledJoinControl.bind(this),
     };
+
+    this.meetingControls[ROSTER_CONTROL] = {
+      ID: ROSTER_CONTROL,
+      action: this.toggleRoster.bind(this),
+      display: this.rosterControl.bind(this),
+    };
   }
 
   /**
@@ -170,6 +177,7 @@ export default class MeetingsJSONAdapter extends MeetingsAdapter {
           remoteAudio: null,
           remoteVideo: null,
           remoteShare: null,
+          showRoster: null,
         });
       } else if (this.datasource[ID]) {
         const meeting = this.datasource[ID];
@@ -177,6 +185,11 @@ export default class MeetingsJSONAdapter extends MeetingsAdapter {
         // Add a video stream as if it were a remote meeting
         if (meeting.remoteVideo instanceof MediaStream) {
           meeting.remoteVideo = this.getVideoStream();
+        }
+
+        // Add a share stream as if it were a remote sharing
+        if (meeting.remoteShare instanceof MediaStream) {
+          meeting.remoteShare = this.getShareStream();
         }
 
         observer.next(meeting);
@@ -188,15 +201,7 @@ export default class MeetingsJSONAdapter extends MeetingsAdapter {
     });
 
     // Send updates on the meeting when an action is triggered
-    const shareEvents$ = fromEvent(document, SHARE_CONTROL).pipe(
-      tap(() => {
-        const meeting = this.datasource[ID];
-
-        // Use local screen share to fake a remote user's screen sharing
-        meeting.remoteShare = meeting.localShare;
-      }),
-    );
-
+    const shareEvents$ = fromEvent(document, SHARE_CONTROL);
     const audioEvents$ = fromEvent(document, MUTE_AUDIO_CONTROL);
     const videoEvents$ = fromEvent(document, MUTE_VIDEO_CONTROL);
     const joinEvents$ = fromEvent(document, JOIN_CONTROL);
@@ -204,8 +209,16 @@ export default class MeetingsJSONAdapter extends MeetingsAdapter {
       filter((event) => event.detail.ID === ID),
       tap(() => end$.next(`Meeting "${ID}" has completed.`)),
     );
+    const rosterEvents$ = fromEvent(document, ROSTER_CONTROL);
 
-    const events$ = merge(audioEvents$, videoEvents$, joinEvents$, leaveEvents$, shareEvents$).pipe(
+    const events$ = merge(
+      audioEvents$,
+      videoEvents$,
+      joinEvents$,
+      leaveEvents$,
+      shareEvents$,
+      rosterEvents$,
+    ).pipe(
       filter((event) => event.detail.ID === ID),
       // Make a copy of the meeting to treat it as if were immutable
       map((event) => ({...event.detail})),
@@ -313,8 +326,8 @@ export default class MeetingsJSONAdapter extends MeetingsAdapter {
    * @private
    */
   // eslint-disable-next-line class-methods-use-this
-  getVideoStream() {
-    const video = document.getElementById('remote-video');
+  getStreamFromElement(id) {
+    const video = document.getElementById(id);
     let stream = new MediaStream();
 
     if (video && video.captureStream) {
@@ -324,6 +337,28 @@ export default class MeetingsJSONAdapter extends MeetingsAdapter {
     }
 
     return stream;
+  }
+
+  /**
+   * Returns a promise to a MediaStream object that captures the contents of a video.
+   *
+   * @returns {Promise.<MediaStream>} Media stream that captures display
+   * @private
+   */
+  // eslint-disable-next-line class-methods-use-this
+  getVideoStream() {
+    return this.getStreamFromElement('remote-video');
+  }
+
+  /**
+   * Returns a promise to a MediaStream object that captures the contents of a screen share
+   *
+   * @returns {Promise.<MediaStream>} Media stream that captures display
+   * @private
+   */
+  // eslint-disable-next-line class-methods-use-this
+  getShareStream() {
+    return this.getStreamFromElement('remote-share');
   }
 
   /**
@@ -378,13 +413,13 @@ export default class MeetingsJSONAdapter extends MeetingsAdapter {
   muteAudioControl(ID) {
     const unmuted = {
       ID: MUTE_AUDIO_CONTROL,
-      icon: 'microphone-muted',
+      icon: 'microphone-muted_28',
       tooltip: 'Mute',
       state: MeetingControlState.INACTIVE,
     };
     const muted = {
       ID: MUTE_AUDIO_CONTROL,
-      icon: 'microphone-muted',
+      icon: 'microphone-muted_28',
       tooltip: 'Unmute',
       state: MeetingControlState.ACTIVE,
     };
@@ -419,14 +454,14 @@ export default class MeetingsJSONAdapter extends MeetingsAdapter {
   muteVideoControl(ID) {
     const muted = {
       ID: MUTE_VIDEO_CONTROL,
-      icon: 'camera-muted',
+      icon: 'camera-muted_28',
       tooltip: 'Start video',
       state: MeetingControlState.ACTIVE,
       text: null,
     };
     const unmuted = {
       ID: MUTE_VIDEO_CONTROL,
-      icon: 'camera-muted',
+      icon: 'camera-muted_28',
       tooltip: 'Stop video',
       state: MeetingControlState.INACTIVE,
       text: null,
@@ -483,7 +518,7 @@ export default class MeetingsJSONAdapter extends MeetingsAdapter {
     return Observable.create((observer) => {
       observer.next({
         ID: LEAVE_CONTROL,
-        icon: 'cancel',
+        icon: 'cancel_28',
         tooltip: 'Leave',
         state: MeetingControlState.ACTIVE,
       });
@@ -526,12 +561,12 @@ export default class MeetingsJSONAdapter extends MeetingsAdapter {
       meeting.localShare = null;
     } else {
       meeting.localShare = await this.getDisplayStream();
+      meeting.remoteShare = null;
 
       if (meeting.localShare) {
         // Handle browser's built-in stop Button
         meeting.localShare.getVideoTracks()[0].onended = () => {
           meeting.localShare = null;
-          meeting.remoteShare = null;
           document.dispatchEvent(new CustomEvent(SHARE_CONTROL, {detail: meeting}));
         };
       }
@@ -550,13 +585,13 @@ export default class MeetingsJSONAdapter extends MeetingsAdapter {
   shareControl(ID) {
     const inactive = {
       ID: SHARE_CONTROL,
-      icon: 'share',
+      icon: 'share-screen-presence-stroke_26',
       tooltip: 'Start Sharing',
       state: MeetingControlState.INACTIVE,
     };
     const active = {
       ID: SHARE_CONTROL,
-      icon: 'share',
+      icon: 'share-screen-presence-stroke_26',
       tooltip: 'Stop Sharing',
       state: MeetingControlState.ACTIVE,
     };
@@ -582,6 +617,64 @@ export default class MeetingsJSONAdapter extends MeetingsAdapter {
   }
 
   /**
+   * Handles the toggle of the roster.
+   *
+   * @param {string} ID ID of the meeting for which to update display
+   * @private
+   */
+  async toggleRoster(ID) {
+    const meeting = this.datasource[ID];
+
+    meeting.showRoster = !meeting.showRoster;
+
+    document.dispatchEvent(new CustomEvent(ROSTER_CONTROL, {detail: meeting}));
+  }
+
+  /**
+   * Returns an observable that emits the display data of a roster control.
+   *
+   * @private
+   * @param {string} ID ID of the meeting to toggle roster
+   * @returns {Observable.<MeetingControlDisplay>} Observable stream that emits display data of the roster control
+   */
+  rosterControl(ID) {
+    const active = {
+      ID: ROSTER_CONTROL,
+      icon: 'participant-list_28',
+      tooltip: 'Hide participants panel',
+      state: MeetingControlState.ACTIVE,
+      text: 'Participants',
+    };
+    const inactive = {
+      ID: ROSTER_CONTROL,
+      icon: 'participant-list_28',
+      tooltip: 'Show participants panel',
+      state: MeetingControlState.INACTIVE,
+      text: 'Participants',
+    };
+
+    const initialState$ = new Observable((observer) => {
+      const meeting = this.datasource[ID];
+
+      if (meeting) {
+        const initialControl = meeting.showRoster ? active : inactive;
+
+        observer.next(initialControl);
+        observer.complete();
+      } else {
+        observer.error(new Error(`Could not find meeting with ID "${ID}" to add roster control`));
+      }
+    });
+
+    const rosterEvent$ = fromEvent(document, ROSTER_CONTROL).pipe(
+      filter((event) => event.detail.ID === ID),
+      map((event) => (event.detail.showRoster ? active : inactive)),
+    );
+
+    return concat(initialState$, rosterEvent$);
+  }
+
+  /**
    * Returns an observable that emits the display data of a disabled meeting control.
    *
    * @returns {Observable.<MeetingControlDisplay>} Observable that emits control display data
@@ -592,7 +685,7 @@ export default class MeetingsJSONAdapter extends MeetingsAdapter {
     return Observable.create((observer) => {
       observer.next({
         ID: JOIN_CONTROL,
-        icon: 'microphone',
+        icon: 'microphone_28',
         tooltip: 'Mute disabled',
         state: MeetingControlState.DISABLED,
       });