+++ src/adapters/MeetingsJSONAdapter.test.js
@@ -27,6 +27,7 @@ describe('Meetings JSON Adapter', () => {
       remoteAudio: null,
       remoteVideo: null,
       remoteShare: null,
+      showRoster: null,
     };
   });
 
@@ -283,7 +284,7 @@ describe('Meetings JSON Adapter', () => {
       meetingsJSONAdapter.leaveControl().subscribe((display) => {
         expect(display).toMatchObject({
           ID: 'leave-meeting',
-          icon: 'cancel',
+          icon: 'cancel_28',
           tooltip: 'Leave',
           state: 'active',
         });
@@ -305,7 +306,7 @@ describe('Meetings JSON Adapter', () => {
         .subscribe((display) => {
           expect(display).toMatchObject({
             ID: 'mute-audio',
-            icon: 'microphone-muted',
+            icon: 'microphone-muted_28',
             tooltip: 'Mute',
             state: 'inactive',
           });
@@ -321,7 +322,7 @@ describe('Meetings JSON Adapter', () => {
         .subscribe((display) => {
           expect(display).toMatchObject({
             ID: 'mute-audio',
-            icon: 'microphone-muted',
+            icon: 'microphone-muted_28',
             tooltip: 'Unmute',
             state: 'active',
           });
@@ -336,7 +337,7 @@ describe('Meetings JSON Adapter', () => {
         .subscribe((display) => {
           expect(display).toMatchObject({
             ID: 'mute-audio',
-            icon: 'microphone-muted',
+            icon: 'microphone-muted_28',
             tooltip: 'Unmute',
             state: 'active',
           });
@@ -371,7 +372,7 @@ describe('Meetings JSON Adapter', () => {
         .subscribe((display) => {
           expect(display).toMatchObject({
             ID: 'mute-video',
-            icon: 'camera-muted',
+            icon: 'camera-muted_28',
             tooltip: 'Stop video',
             state: 'inactive',
           });
@@ -387,7 +388,7 @@ describe('Meetings JSON Adapter', () => {
         .subscribe((display) => {
           expect(display).toMatchObject({
             ID: 'mute-video',
-            icon: 'camera-muted',
+            icon: 'camera-muted_28',
             tooltip: 'Start video',
             state: 'active',
           });
@@ -402,7 +403,7 @@ describe('Meetings JSON Adapter', () => {
         .subscribe((display) => {
           expect(display).toMatchObject({
             ID: 'mute-video',
-            icon: 'camera-muted',
+            icon: 'camera-muted_28',
             tooltip: 'Start video',
             state: 'active',
           });
@@ -418,6 +419,71 @@ describe('Meetings JSON Adapter', () => {
         () => {},
         (error) => {
           expect(error.message).toEqual('Could not find meeting with ID "invalid"');
+          done();
+        },
+      );
+    });
+  });
+
+  describe('toggleRoster()', () => {
+    let dispatchSpy;
+
+    beforeEach(() => {
+      dispatchSpy = jest.spyOn(document, 'dispatchEvent');
+    });
+
+    afterEach(() => {
+      dispatchSpy.mockRestore();
+    });
+
+    test('dispatches a "member-roster" event', async () => {
+      await meetingsJSONAdapter.toggleRoster(meetingID);
+      expect(dispatchSpy).toHaveBeenCalled();
+    });
+  });
+
+  describe('rosterControl()', () => {
+    test('returns an observable', () => {
+      expect(isObservable(meetingsJSONAdapter.rosterControl())).toBeTruthy();
+    });
+
+    test('emits inactive roster icon control if roster is not displayed', (done) => {
+      meetingsJSONAdapter
+        .rosterControl(meetingID)
+        .subscribe((display) => {
+          expect(display).toMatchObject({
+            ID: 'member-roster',
+            icon: 'participant-list_28',
+            tooltip: 'Show participants panel',
+            state: 'inactive',
+            text: 'Participants',
+          });
+          done();
+        });
+    });
+
+    test('emits active roster icon control if roster is displayed', (done) => {
+      testMeeting.showRoster = true;
+
+      meetingsJSONAdapter
+        .rosterControl(meetingID)
+        .subscribe((display) => {
+          expect(display).toMatchObject({
+            ID: 'member-roster',
+            icon: 'participant-list_28',
+            tooltip: 'Hide participants panel',
+            state: 'active',
+            text: 'Participants',
+          });
+          done();
+        });
+    });
+
+    test('throws error on invalid meeting ID', (done) => {
+      meetingsJSONAdapter.rosterControl('invalid').subscribe(
+        () => {},
+        (error) => {
+          expect(error.message).toEqual('Could not find meeting with ID "invalid" to add roster control');
           done();
         },
       );