+++ src/components/WebexMemberRoster/WebexMemberRoster.stories.js
@@ -4,6 +4,7 @@ import WebexMemberRoster from './WebexMemberRoster';
 export default {
   title: 'Platform/Webex Member Roster',
   component: WebexMemberRoster,
+  decorators: [(Story) => <div style={{width: '20rem'}}><Story /></div>],
 };
 
 const Template = (args) => <WebexMemberRoster {...args} />;
@@ -11,5 +12,17 @@ const Template = (args) => <WebexMemberRoster {...args} />;
 export const Space = Template.bind({});
 Space.args = {
   destinationID: 'room1',
+  destinationType: 'room',
+};
+
+export const Meeting = Template.bind({});
+Meeting.args = {
+  destinationID: 'meeting1',
+  destinationType: 'meeting',
+};
+
+export const ExternalMember = Template.bind({});
+ExternalMember.args = {
+  destinationID: 'room2',
   destinationType: 'room',
 };